Refactor API views into submodules
